Let Tux bounce off badguys when he's invincible. This let's us bounce off guys like the flying snowball without falling to our deaths. Tux no longer kicks mriceblocks right before picking them up.
SVN-Revision: 3456

+++ src/badguy/badguy.cpp
@@ -167,11 +167,6 @@ BadGuy::collision_solid(GameObject& , const CollisionHit& )
 HitResponse
 BadGuy::collision_player(Player& player, const CollisionHit& )
 {
-  if(player.is_invincible()) {
-    kill_fall();
-    return ABORT_MOVE;
-  }
-
   /*
   printf("PlayerHit: GT %3.1f PM: %3.1f %3.1f BM: %3.1f %3.1f Hit: %3.1f %3.1f\n",
           game_time,
@@ -185,6 +180,11 @@ BadGuy::collision_player(Player& player, const CollisionHit& )
     // if it's not possible to squish us, then this will hurt
     if(collision_squished(player))
       return ABORT_MOVE;
+  }
+
+  if(player.is_invincible()) {
+    kill_fall();
+    return ABORT_MOVE;
   }
 
   player.kill(false);
